refactor: SponsorManagement to SponsorProfileForm (#862)
* refactor: SponsorManagement to SponsorProfileForm

* pass through comments SponsorProfileform
